+++ packages/fiori/src/NotificationListItem.ts
@@ -8,7 +8,6 @@ import slot from "@ui5/webcomponents-base/dist/decorators/slot.js";
 import event from "@ui5/webcomponents-base/dist/decorators/event.js";
 import ResizeHandler from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
 import type { ResizeObserverCallback } from "@ui5/webcomponents-base/dist/delegate/ResizeHandler.js";
-import { getEventMark } from "@ui5/webcomponents-base/dist/MarkedEvents.js";
 import Button from "@ui5/webcomponents/dist/Button.js";
 import BusyIndicator from "@ui5/webcomponents/dist/BusyIndicator.js";
 import Tag from "@ui5/webcomponents/dist/Tag.js";
@@ -492,8 +491,8 @@ class NotificationListItem extends NotificationListItemBase {
 	/**
 	 * Event handlers
 	 */
-	_onclick(e: MouseEvent) {
-		this.fireItemPress(e);
+	_onclick() {
+		this.fireItemPress();
 	}
 
 	_onShowMoreClick(e: MouseEvent) {
@@ -550,7 +549,7 @@ class NotificationListItem extends NotificationListItemBase {
 
 		const space = isSpace(e);
 
-		if (space && getEventMark(e) === "link") {
+		if (space && this.getFocusDomRef()!.matches(":has(:focus-within)")) {
 			this._onShowMoreClick(e as unknown as MouseEvent);
 			return;
 		}
@@ -592,8 +591,8 @@ class NotificationListItem extends NotificationListItemBase {
 	/**
 	 * Private
 	 */
-	fireItemPress(e: Event) {
-		if (getEventMark(e) === "button" || getEventMark(e) === "link") {
+	fireItemPress() {
+		if (this.getFocusDomRef()!.matches(":has(:focus-within)")) {
 			return;
 		}