+++ packages/main/src/SplitButton.ts
@@ -19,7 +19,6 @@ import {
 import AriaHasPopup from "@ui5/webcomponents-base/dist/types/AriaHasPopup.js";
 import i18n from "@ui5/webcomponents-base/dist/decorators/i18n.js";
 import type I18nBundle from "@ui5/webcomponents-base/dist/i18nBundle.js";
-import { getEventMark } from "@ui5/webcomponents-base/dist/MarkedEvents.js";
 import litRender from "@ui5/webcomponents-base/dist/renderer/LitRenderer.js";
 import "@ui5/webcomponents-icons/dist/slim-arrow-down.js";
 import type ButtonDesign from "./types/ButtonDesign.js";
@@ -233,8 +232,8 @@ class SplitButton extends UI5Element {
 		this._fireClick(e);
 	}
 
-	_onFocusOut(e: FocusEvent) {
-		if (this.disabled || getEventMark(e)) {
+	_onFocusOut() {
+		if (this.disabled || this.getFocusDomRef()!.matches(":has(:focus-within)")) {
 			return;
 		}
 
@@ -242,8 +241,8 @@ class SplitButton extends UI5Element {
 		this._setTabIndexValue();
 	}
 
-	_onFocusIn(e: FocusEvent) {
-		if (this.disabled || getEventMark(e)) {
+	_onFocusIn() {
+		if (this.disabled || this.getFocusDomRef()!.matches(":has(:focus-within)")) {
 			return;
 		}
 		this._shiftOrEscapePressed = false;