+++ packages/website/docs/_samples/fiori/NavigationLayout/Basic/main.js
@@ -1,6 +1,10 @@
 import "@ui5/webcomponents/dist/Button.js";
+import "@ui5/webcomponents/dist/Text.js";
+import "@ui5/webcomponents/dist/Title.js";
 
+import "@ui5/webcomponents-fiori/dist/NavigationLayout.js";
 import "@ui5/webcomponents-fiori/dist/SideNavigation.js";
+import "@ui5/webcomponents-fiori/dist/SideNavigationGroup.js";
 import "@ui5/webcomponents-fiori/dist/SideNavigationItem.js";
 import "@ui5/webcomponents-fiori/dist/SideNavigationSubItem.js";
 import "@ui5/webcomponents-fiori/dist/ShellBar.js";
@@ -17,8 +21,22 @@ import "@ui5/webcomponents-icons/dist/background.js";
 import "@ui5/webcomponents-icons/dist/activity-assigned-to-goal.js";
 import "@ui5/webcomponents-icons/dist/action-settings.js";
 import "@ui5/webcomponents-icons/dist/chain-link.js";
+import "@ui5/webcomponents-icons/dist/document-text.js";
+import "@ui5/webcomponents-icons/dist/compare.js";
+import "@ui5/webcomponents-icons/dist/locked.js";
 
-var sidenav = document.querySelector("ui5-side-navigation");
-document.getElementById("toggle").addEventListener("click", () => {
-    sidenav.toggleAttribute("collapsed");
+document.querySelector("#startButton").addEventListener("click", function (event) {
+	nl1.sideCollapsed = !nl1.sideCollapsed;
+});
+
+document.querySelector("#sn1").addEventListener("selection-change", function (event) {
+	if (event.detail.item.getAttribute("target")) {
+		return;
+	}
+
+	const contentItems = document.querySelectorAll(".contentItem");
+	contentItems.forEach(item => {
+		item.classList.remove("contentItemVisible");
+	});
+	document.getElementById(event.detail.item.getAttribute("href").replace("#", "")).classList.add("contentItemVisible");
 });